--- a/pandas/tests/reshape/merge/test_merge_asof.py
+++ b/pandas/tests/reshape/merge/test_merge_asof.py
@@ -288,8 +288,7 @@ class TestAsOfMerge:
         expected.loc[expected.ticker == "MSFT", ["bid", "ask"]] = np.nan
         tm.assert_frame_equal(result, expected)
 
-    @pytest.mark.parametrize("dtype_backend", [None, "numpy_nullable", "pyarrow"])
-    def test_multiby(self, dtype_backend):
+    def test_multiby(self):
         # GH13936
         trades = pd.DataFrame(
             {
@@ -309,8 +308,6 @@ class TestAsOfMerge:
             },
             columns=["time", "ticker", "exch", "price", "quantity"],
         )
-        if dtype_backend is not None:
-            trades = trades.convert_dtypes(dtype_backend=dtype_backend)
 
         quotes = pd.DataFrame(
             {
@@ -331,8 +328,6 @@ class TestAsOfMerge:
             },
             columns=["time", "ticker", "exch", "bid", "ask"],
         )
-        if dtype_backend is not None:
-            quotes = quotes.convert_dtypes(dtype_backend=dtype_backend)
 
         expected = pd.DataFrame(
             {
@@ -354,13 +349,12 @@ class TestAsOfMerge:
             },
             columns=["time", "ticker", "exch", "price", "quantity", "bid", "ask"],
         )
-        if dtype_backend is not None:
-            expected = expected.convert_dtypes(dtype_backend=dtype_backend)
 
         result = merge_asof(trades, quotes, on="time", by=["ticker", "exch"])
         tm.assert_frame_equal(result, expected)
 
-    def test_multiby_heterogeneous_types(self):
+    @pytest.mark.parametrize("dtype", ["object", "string"])
+    def test_multiby_heterogeneous_types(self, dtype):
         # GH13936
         trades = pd.DataFrame(
             {
@@ -380,6 +374,7 @@ class TestAsOfMerge:
             },
             columns=["time", "ticker", "exch", "price", "quantity"],
         )
+        trades = trades.astype({"ticker": dtype, "exch": dtype})
 
         quotes = pd.DataFrame(
             {
@@ -400,6 +395,7 @@ class TestAsOfMerge:
             },
             columns=["time", "ticker", "exch", "bid", "ask"],
         )
+        quotes = quotes.astype({"ticker": dtype, "exch": dtype})
 
         expected = pd.DataFrame(
             {
@@ -421,6 +417,7 @@ class TestAsOfMerge:
             },
             columns=["time", "ticker", "exch", "price", "quantity", "bid", "ask"],
         )
+        expected = expected.astype({"ticker": dtype, "exch": dtype})
 
         result = merge_asof(trades, quotes, on="time", by=["ticker", "exch"])
         tm.assert_frame_equal(result, expected)